requirements.txt
torch==1.2.0
torchvision==0.4.0

numpy==1.14.5
scipy==1.2.0
scikit_learn==0.20.2
matplotlib~=2.2.0
tqdm==4.23.4

# mxnet==1.4.0
# mxnet-cu92==1.4.0

opencv_python==3.4.4.19
Pillow==5.2.0

bcolz==1.2.0
easydict==1.7

tensorboardX==1.6

protobuf==3.15.0
pybind11[global]

onnx==1.5.0
